requirements.txt
pyttsx3~=2.90
SpeechRecognition
pyaudio
pystray~=0.19.4
Pillow~=9.5.0